// final exit project/Angular/Frontend/src/app/models/alltickets.ts
export class Alltickets {
  id: string;
  requestType = '';
  priority = '';
  travelCity = '';
  fromLocationCity = '';
  travelStartDate: Date;
  travelEndDate: Date;
  passportNumber = '';
  projectName = '';
  expenseBorneBy = '';
  travelApproverName: string;
  expectedDurationOfTravel: string;
  upperBound: string;
  anyAdditionalDetails = '';
  comments = '';
  status = '';
  files = null;
  user = null;
}